# bookstore-be/resources/api.specs.json
{
  "openapi":"3.0.1",
  "info":{
    "title":"OpenAPI definition",
    "version":"v0"
  },
  "servers":[
    {
      "url":"http://localhost:8080/",
      "description":"Generated server url"
    }
  ],
  "paths":{
    "/bookstore/add":{
      "post":{
        "tags":[
          "bookstore-controller"
        ],
        "operationId":"insertOrder",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/BuyBookDTO"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/MonoResponseEntityOrderBookDTO"
                }
              }
            }
          }
        }
      }
    },
    "/bookstore/detail/{id}":{
      "get":{
        "tags":[
          "bookstore-controller"
        ],
        "operationId":"getBookDetail",
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"string"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/MonoResponseEntityBooksDTO"
                }
              }
            }
          }
        }
      }
    },
    "/bookstore/list":{
      "get":{
        "tags":[
          "bookstore-controller"
        ],
        "operationId":"getBookstoreList",
        "parameters":[
          {
            "name":"parameters",
            "in":"query",
            "required":true,
            "schema":{
              "$ref":"#/components/schemas/SearchParameters"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/MonoResponseEntityListBooksDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components":{
    "schemas":{
      "BuyBookDTO":{
        "type":"object",
        "properties":{
          "service":{
            "type":"string"
          },
          "cost":{
            "type":"number",
            "format":"double"
          }
        }
      },
      "MonoResponseEntityOrderBookDTO":{
        "type":"object"
      },
      "MonoResponseEntityBooksDTO":{
        "type":"object"
      },
      "SearchParameters":{
        "type":"object",
        "properties":{
          "keyword":{
            "type":"string"
          },
          "searchType":{
            "type":"string"
          }
        }
      },
      "MonoResponseEntityListBooksDTO":{
        "type":"object"
      }
    }
  }
}