# bookstore-be/src/main/java/com/augen/bookstore/dto/BuyBookDTO.java
package com.augen.bookstore.dto;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@JsonIgnoreProperties(ignoreUnknown = true)
public class BuyBookDTO {
    private String service;
    private Double cost;
}